Revamped signin/signup UI. Added email form validation.
