+++ src/Client/components/Sign_Up/Sign_Up.js
@@ -1,25 +1,109 @@
 import React, {Component} from 'react';
 import ReactDOM from 'react-dom';
 import { reduxForm } from 'redux-form';
+import  {Link} from 'react-router';
 import * as actions from '../../actions';
 //Custom components
-import Banner from '../../Banner';
-import Sign_Up_Box from './Sign_Up_Box';
 
-export default class Sign_Up extends Component {
+class Sign_Up extends Component {
+
+  handleFormSubmit({name, email, password}){
+    this.props.signupUser({name, email, password});
+  }
+
+  renderAlert(){
+    if(this.props.errorMessage){
+      return(
+        <div className="alert alert-danger">
+          <strong>Oops!</strong> {this.props.errorMessage}
+        </div>
+      );
+    }
+  }
+
   render(){
+    const {handleSubmit, fields: {name, email, password, confirmedPassword}} = this.props;
     return(
-    <div>
-      <Banner/>
-      <div className="panel panel-warning">
-        <div className="panel-heading">
-          <h3 className="panel-title">SIGN UP</h3>
+      <div>
+        <div className="app-name">
+          <h1>ENGAGER</h1>
+          <h4>CREATE AN ACCOUNT</h4>
         </div>
-        <div className="panel-body">
-          <Sign_Up_Box/>
-        </div>
+        <form className="box-form" onSubmit={handleSubmit(this.handleFormSubmit.bind(this))}>
+          <fieldset className="form-group">
+            <input type="text" className="form-control"
+              placeholder="Name"
+             {...name}/>
+             {name.touched && name.error && <div className="text-error">{name.error}</div>}
+           </fieldset>
+          <fieldset className="form-group">
+          <input type="text" className="form-control"
+            placeholder="Email e.g. example@octoconsulting.com"
+            pattern="[A-Za-z0-9\.]+@octoconsulting\.com"
+            title="Only A Valid Octo Email Is Allowed"
+            {...email}/>
+            {email.touched && email.error && <div className="text-error">{email.error}</div>}
+          </fieldset>
+          <fieldset className="form-group">
+            <input type="password" className="form-control" placeholder="Password"
+              {...password}/>
+              {password.touched && password.error && <div className="text-error">{password.error}</div>}
+            </fieldset>
+        <fieldset className="form-group">
+            <input type="password" className="form-control" placeholder="Confirm Password"
+            {...confirmedPassword}/>
+            <div className="text-error">
+              {confirmedPassword.touched ? confirmedPassword.error : ''}
+            </div>
+            </fieldset>
+            {this.renderAlert()}
+            <div className="button_pos">
+              <button type="submit" className="btn btn-primary">REGISTER </button>
+            </div>
+            <div className="button_pos">
+              <Link to='/SignIn' className="btn btn-primary">CANCEL</Link>
+            </div>
+        </form>
       </div>
-    </div>
     );
   }
 }
+
+function validate(values){
+  const errors ={};
+  if (values.password != values.confirmedPassword){
+    errors.confirmedPassword = "Passwords must match";
+  }
+
+  if (!values.name){
+    errors.name = "Please enter a name";
+  }
+  if (!values.email){
+    errors.email = "Please enter a valid Octo email";
+  }
+  if (!values.password){
+    errors.password = "Please enter a password";
+  }
+  if (!values.confirmedPassword){
+    errors.confirmedPassword = "Please enter your password a second time";
+  }
+
+  return errors;
+}
+
+function mapStateToProps(state){
+  return {
+    errorMessage: state.auth.error
+  }
+}
+
+export default reduxForm({
+  form: 'signup',
+  fields: [
+    'name',
+    'email',
+    'password',
+    'confirmedPassword'
+  ],
+  validate
+},mapStateToProps, actions)(Sign_Up);
